feat: move table_test_fixture to test_utils crate
