+++ src/query/service/Cargo.toml
@@ -16,7 +16,7 @@ default = ["simd"]
 simd = ["common-arrow/simd"]
 
 tokio-console = ["common-tracing/console", "common-base/tracing"]
-memory-profiling = ["common-base/memory-profiling", "common-http/memory-profiling", "tempfile"]
+memory-profiling = ["common-base/memory-profiling", "common-http/memory-profiling"]
 storage-hdfs = ["opendal/services-hdfs", "common-storage/storage-hdfs"]
 hive = [
     "storage-hdfs",
@@ -145,7 +145,7 @@ serde_json = { workspace = true }
 serde_urlencoded = "0.7.1"
 socket2 = "0.4.7"
 strength_reduce = "0.2.4"
-tempfile = { version = "3.4.0", optional = true }
+tempfile = "3.4.0"
 time = "0.3.14"
 tokio-stream = { version = "0.1.10", features = ["net"] }
 tonic = { workspace = true }
@@ -153,6 +153,7 @@ tracing = "0.1.36"
 typetag = "0.2.3"
 unicode-segmentation = "1.10.1"
 uuid = { version = "1.1.2", features = ["serde", "v4"] }
+walkdir = "2.3.2"
 
 [dev-dependencies]
 arrow-cast = { version = "37.0.0", features = ["prettyprint"] }
