+++ src/query/service/src/lib.rs
@@ -46,6 +46,7 @@ pub mod servers;
 pub mod sessions;
 pub mod stream;
 pub mod table_functions;
+pub mod test_utils;
 
 mod global_services;
