+++ src/query/service/tests/it/api/http/cluster.rs
@@ -16,6 +16,7 @@ use common_base::base::tokio;
 use common_exception::Result;
 use common_meta_types::NodeInfo;
 use databend_query::api::http::v1::cluster::*;
+use databend_query::test_utils::TestGlobalServices;
 use poem::get;
 use poem::http::header;
 use poem::http::Method;
@@ -26,11 +27,11 @@ use poem::Request;
 use poem::Route;
 use pretty_assertions::assert_eq;
 
-use crate::tests::TestGlobalServices;
-
 #[tokio::test(flavor = "multi_thread")]
 async fn test_cluster() -> Result<()> {
-    let _guard = TestGlobalServices::setup(crate::tests::ConfigBuilder::create().build()).await?;
+    let _guard =
+        TestGlobalServices::setup(databend_query::test_utils::ConfigBuilder::create().build())
+            .await?;
     let cluster_router = Route::new().at("/v1/cluster/list", get(cluster_list_handler));
 
     // List Node
