+++ src/query/service/tests/it/api/http/config.rs
@@ -14,6 +14,7 @@
 
 use common_base::base::tokio;
 use databend_query::api::http::v1::config::config_handler;
+use databend_query::test_utils::TestGlobalServices;
 use poem::get;
 use poem::http::Method;
 use poem::http::StatusCode;
@@ -23,11 +24,11 @@ use poem::Request;
 use poem::Route;
 use pretty_assertions::assert_eq; // for `app.oneshot()`
 
-use crate::tests::TestGlobalServices;
-
 #[tokio::test(flavor = "multi_thread")]
 async fn test_config() -> common_exception::Result<()> {
-    let _guard = TestGlobalServices::setup(crate::tests::ConfigBuilder::create().build()).await?;
+    let _guard =
+        TestGlobalServices::setup(databend_query::test_utils::ConfigBuilder::create().build())
+            .await?;
     let cluster_router = Route::new().at("/v1/config", get(config_handler));
 
     let response = cluster_router
