+++ src/query/service/tests/it/sessions/query_ctx.rs
@@ -33,7 +33,7 @@ async fn test_get_storage_accessor_s3() -> Result<()> {
         .mount(&mock_server)
         .await;
 
-    let mut conf = crate::tests::ConfigBuilder::create().config();
+    let mut conf = databend_query::test_utils::ConfigBuilder::create().config();
 
     conf.storage.params = StorageParams::S3(StorageS3Config {
         region: "us-east-2".to_string(),
@@ -43,7 +43,8 @@ async fn test_get_storage_accessor_s3() -> Result<()> {
         ..Default::default()
     });
 
-    let (_guard, qctx) = crate::tests::create_query_context_with_config(conf, None).await?;
+    let (_guard, qctx) =
+        databend_query::test_utils::create_query_context_with_config(conf, None).await?;
 
     let _ = qctx.get_data_operator()?;
 
@@ -52,13 +53,14 @@ async fn test_get_storage_accessor_s3() -> Result<()> {
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_get_storage_accessor_fs() -> Result<()> {
-    let mut conf = crate::tests::ConfigBuilder::create().config();
+    let mut conf = databend_query::test_utils::ConfigBuilder::create().config();
 
     conf.storage.params = StorageParams::Fs(StorageFsConfig {
         root: "/tmp".to_string(),
     });
 
-    let (_guard, qctx) = crate::tests::create_query_context_with_config(conf, None).await?;
+    let (_guard, qctx) =
+        databend_query::test_utils::create_query_context_with_config(conf, None).await?;
 
     let _ = qctx.get_data_operator()?;
