+++ src/query/service/tests/it/sessions/session_setting.rs
@@ -16,12 +16,13 @@ use common_base::base::tokio;
 use common_exception::Result;
 use databend_query::sessions::SessionManager;
 use databend_query::sessions::SessionType;
-
-use crate::tests::TestGlobalServices;
+use databend_query::test_utils::TestGlobalServices;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_session_setting() -> Result<()> {
-    let _guard = TestGlobalServices::setup(crate::tests::ConfigBuilder::create().build()).await?;
+    let _guard =
+        TestGlobalServices::setup(databend_query::test_utils::ConfigBuilder::create().build())
+            .await?;
     let session = SessionManager::instance()
         .create_session(SessionType::Dummy)
         .await?;
@@ -41,7 +42,7 @@ async fn test_session_setting() -> Result<()> {
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_session_setting_override() -> Result<()> {
     let _guard = TestGlobalServices::setup(
-        crate::tests::ConfigBuilder::create()
+        databend_query::test_utils::ConfigBuilder::create()
             .max_storage_io_requests(1000)
             .parquet_fast_read_bytes(1000000)
             .build(),
