+++ src/query/service/tests/it/storages/fuse/mod.rs
@@ -22,5 +22,4 @@ mod pruning;
 mod statistics;
 mod table;
 mod table_functions;
-mod table_test_fixture;
 mod utils;
