+++ src/query/service/tests/it/storages/fuse/operations/commit.rs
@@ -83,6 +83,8 @@ use common_storages_fuse::FuseTable;
 use common_storages_fuse::FUSE_TBL_SNAPSHOT_PREFIX;
 use dashmap::DashMap;
 use databend_query::sessions::QueryContext;
+use databend_query::test_utils::table_test_fixture::execute_query;
+use databend_query::test_utils::table_test_fixture::TestFixture;
 use futures::TryStreamExt;
 use rand::thread_rng;
 use rand::Rng;
@@ -92,8 +94,6 @@ use walkdir::WalkDir;
 
 use crate::storages::fuse::block_writer::BlockWriter;
 use crate::storages::fuse::operations::mutation::CompactSegmentTestFixture;
-use crate::storages::fuse::table_test_fixture::execute_query;
-use crate::storages::fuse::table_test_fixture::TestFixture;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_fuse_occ_retry() -> Result<()> {
