+++ src/query/service/tests/it/storages/null.rs
@@ -28,7 +28,7 @@ use futures::TryStreamExt;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_null_table() -> Result<()> {
-    let (_guard, ctx) = crate::tests::create_query_context().await?;
+    let (_guard, ctx) = databend_query::test_utils::create_query_context().await?;
 
     let table = NullTable::try_create(TableInfo {
         desc: "'default'.'a'".into(),
